% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDownloaders.R
\docType{methods}
\name{loadMIGSAdata}
\alias{loadMIGSAdata}
\alias{loadMIGSAdata,character-method}
\title{Functions to download different RDatas}
\usage{
loadMIGSAdata(dataName)

\S4method{loadMIGSAdata}{character}(dataName)
}
\arguments{
\item{dataName}{character name of the data to download, one of
\itemize{
\item "tcgaMAdata" (~ 8.8 MB) tcga microarray expresion matrix and subtypes.
\item "tcgaRNAseqData" (~ 7.8 MB) tcga RNAseq expresion matrix and subtypes.
\item "pbcmcData" (~ 51 MB) pbcmc microarray expresion matrices and subtypes.
\item "bcMigsaRes" (~ 15 MB) breast cancer MIGSA results.
}}
}
\value{
list with downloaded data.
}
\description{
Different functions to download datasets used in MIGSA's vignette.
}
\examples{
## Lets load tcga MicroArray data.
tcgaMAdata <- loadMIGSAdata("tcgaMAdata");

## Lets load tcga RNAseq data.
tcgaRNAseqData <- loadMIGSAdata("tcgaRNAseqData");

## Lets load pbcmc microarray datasets.
pbcmcData <- loadMIGSAdata("pbcmcData");

## Lets load breast cancer MIGSA results.
bcMigsaRes <- loadMIGSAdata("bcMigsaRes");

}

