% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitOptions-class.R
\docType{methods}
\name{FitOptions-class}
\alias{FitOptions-class}
\title{FitOptions S4 class implementation in R}
\description{
This S4 class contains the parameters to provide for model fitting.
See \link{FitOptions} to create one FitOptions object.
}

