% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geneset-class.R
\docType{methods}
\name{Geneset-class}
\alias{Geneset}
\alias{Geneset-class}
\title{Geneset S4 class implementation in R}
\description{
This S4 class represents an individual gene set.
}
\section{Slots}{

\describe{
\item{\code{id}}{character containing the id of this gene set.}

\item{\code{name}}{character containing the name of this gene set (can be 
empty).}

\item{\code{genes}}{character vector of the genes conforming this gene set 
(must contain at least one gene).}
}}
\examples{
## Lets create manually a gene set
fakeGenes <- c("8818", "2406", "64176");
myGs <- Geneset(id="fakeId", name="fakeName", genes=fakeGenes);

}

