% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets-class.R
\docType{methods}
\name{Genesets-class}
\alias{Genesets}
\alias{Genesets-class}
\title{Genesets S4 class implementation in R}
\description{
This S4 class represents a collection of gene sets.
}
\section{Slots}{

\describe{
\item{\code{name}}{character containing the 
name of this collection of gene sets.}

\item{\code{gene_sets}}{list of Geneset, containing each gene set (can not have 
repeated Geneset ids).}

\item{\code{is_GO}}{logical indicating wether this sets are from Gene 
Ontology (for further analyses and plots). If true, then each Geneset id 
must be a GO id.}
}}
\examples{
## Lets create manually a gene sets containing three gene set.
## First lets create manually three gene set.
myGs1 <- Geneset(id="fakeId1", name="fakeName1", genes=as.character(1:10));
myGs2 <- Geneset(id="fakeId2", name="fakeName2", genes=as.character(7:15));
myGs3 <- Geneset(id="fakeId3", name="fakeName2", genes=as.character(20:28));

## Now we can create our Genesets object.
myGSs <- Genesets(name="myGenesets", gene_sets=list(myGs1, myGs2, myGs3),
is_GO=FALSE);

}

