% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets-downloadEnrichrGeneSets.R
\docType{methods}
\name{downloadEnrichrGeneSets}
\alias{downloadEnrichrGeneSets}
\alias{downloadEnrichrGeneSets,character-method}
\title{Downloads Genesets from enrichr database}
\usage{
downloadEnrichrGeneSets(geneSetNames, ...)

\S4method{downloadEnrichrGeneSets}{character}(geneSetNames,
deleteMultipleEntrez = !FALSE)
}
\arguments{
\item{geneSetNames}{list of characters with the names of the gene sets to 
download. Must be listed at \code{\link{enrichrGeneSets}}.}

\item{...}{not in use.}

\item{deleteMultipleEntrez}{logical indicating if multiple Entrez IDs should 
be deleted or repeated.
Note: Enrichr uses Gene Symbol, if org.Hs.eg translates it into several 
Entrez IDs then if deleteMultipleEntrez == FALSE all Entrez are removed else 
all are included.}
}
\value{
list of Genesets objects.
}
\description{
\code{downloadEnrichrGeneSets} creates a list of Genesets objects 
downloading the specified ones from enrichr website 
(http://amp.pharm.mssm.edu/Enrichr/).
}
\examples{
## Lets download BioCarta gene sets from Enrichr.
## Make sure you use the same names as listed with enrichrGeneSets() .
bioCartaGSs <- downloadEnrichrGeneSets(c("BioCarta_2015"));

}

