% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets-enrichrGeneSets.R
\docType{methods}
\name{enrichrGeneSets}
\alias{enrichrGeneSets}
\alias{enrichrGeneSets,ANY-method}
\title{Lists the gene sets present at enrichr database.}
\usage{
enrichrGeneSets(pattern = ".*")

\S4method{enrichrGeneSets}{ANY}(pattern = ".*")
}
\arguments{
\item{pattern}{character indicating a pattern to filter the database names.}
}
\value{
character with present database names.
}
\description{
\code{enrichrGeneSets} lists the database names present at enrichr.
}
\examples{
## Lets list all the gene sets that can be downloaded from Enichr website.
enrichrGeneSets();

## Now lets list only the gene sets that have KEGG in their names.
enrichrGeneSets("KEGG");

}

