% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets-geneSetsFromFile.R
\docType{methods}
\name{geneSetsFromFile}
\alias{geneSetsFromFile}
\alias{geneSetsFromFile,character-method}
\title{Creates a Genesets from file}
\usage{
geneSetsFromFile(filePath, name, ...)

\S4method{geneSetsFromFile}{character}(filePath, name, is_GO = FALSE)
}
\arguments{
\item{filePath}{character with the path of the file to parse.}

\item{name}{character with the name of this Genesets.}

\item{...}{not in use.}

\item{is_GO}{logical indicating if this Genesets are from the Gene Ontology. 
If true, then each Geneset id must be a GO id.}
}
\value{
Genesets object.
}
\description{
\code{geneSetsFromFile} creates a Genesets object from the data present in 
a file. The file must be a tab separated values file (tsv). Each line will 
parse to a Geneset. First field will be the Geneset id, the second the name 
and the remaining are the genes.
}
\examples{
## Create some fake gene sets in a data.frame to save them in disk and then
## load them (10 gene sets with 20 genes each (it is not neccesary that they
## have the same number of genes).
gsets <- data.frame(
    IDs=paste("set", 1:10), 
    Names=rep("", 10), 
    matrix(paste("gene", 1:(10*20)), nrow=10));

## And save this file as a tab separated file.
write.table(gsets, file="fakeGsets.tsv", sep="\\t",
    col.names=FALSE, row.names=FALSE, quote=FALSE);

## Now lets load this tsv file as a Genesets object.
myGsets <- geneSetsFromFile("fakeGsets.tsv", "fakeGsets");

## And lets delete this tsv file (so we dont have garbage in our disk).
unlink("fakeGsets.tsv");

}

