% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets-getterSetters.R
\docType{methods}
\name{Genesets-getterSetters}
\alias{Genesets-getterSetters}
\alias{geneSets}
\alias{geneSets,Genesets-method}
\alias{geneSets<-}
\alias{geneSets<-,Genesets-method}
\alias{isGO}
\alias{isGO,Genesets-method}
\alias{isGO<-}
\alias{isGO<-,Genesets-method}
\alias{name}
\alias{name,Genesets-method}
\alias{name<-}
\alias{name<-,Genesets-method}
\title{Accessors for Genesets class.}
\usage{
"Genesets-getterSetters"(object)

name(object)

\S4method{name}{Genesets}(object)

name(object) <- value

\S4method{name}{Genesets}(object) <- value

geneSets(object)

\S4method{geneSets}{Genesets}(object)

geneSets(object) <- value

\S4method{geneSets}{Genesets}(object) <- value

isGO(object)

\S4method{isGO}{Genesets}(object)

isGO(object) <- value

\S4method{isGO}{Genesets}(object) <- value
}
\arguments{
\item{object}{Genesets object.}

\item{value}{value to replace in the slot.}
}
\value{
modified object or desired slot.
}
\description{
Getters and setters functions to access Genesets object slots.
}
\examples{
## Lets create manually a gene sets containing three gene set.
## First lets create manually three gene set.
myGs1 <- Geneset(id="fakeId1", name="fakeName1", genes=as.character(1:10));
myGs2 <- Geneset(id="fakeId2", name="fakeName2", genes=as.character(7:15));
myGs3 <- Geneset(id="fakeId3", name="fakeName2", genes=as.character(20:28));

## Now we can create our Genesets object.
myGSs <- Genesets(name="myGenesets", gene_sets=list(myGs1, myGs2, myGs3),
is_GO=FALSE);

## Lets get out Genesets values, and modify its name.
name(myGSs);
name(myGSs) <- "newName";
geneSets(myGSs);
isGO(myGSs);

}

