% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAparams-common.R
\docType{methods}
\name{GSEAparams-common}
\alias{GSEAparams-common}
\alias{summary,GSEAparams-method}
\alias{summary.GSEAparams}
\title{GSEAparams exploratory functions}
\usage{
\method{summary}{GSEAparams}(object, ...)
}
\arguments{
\item{object}{GSEAparams object.}

\item{...}{not in use.}
}
\value{
a summary of the object.
}
\description{
Several R base overwritten functions to manipulate a GSEAparams object.
}
\examples{
gseaParams <- GSEAparams();
summary(gseaParams);

}

