% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGSAinput-class.R
\docType{methods}
\name{IGSAinput-class}
\alias{IGSAinput}
\alias{IGSAinput-class}
\title{IGSAinput S4 class implementation in R}
\description{
This S4 class contains all the necessary inputs to execute a functional 
analysis (SEA and GSEA) on one experiment.
}
\section{Slots}{

\describe{
\item{\code{name}}{character indicating the name of this experiment.}

\item{\code{expr_data}}{ExprData object with the expression data (MicroArray or 
RNAseq).}

\item{\code{use_voom}}{logical indicating wether to use 
\code{\link[limma]{voom}} before fit or not. If using RNAseq data it is 
recommended to set it to TRUE (default: FALSE).}

\item{\code{fit_options}}{FitOptions object with the parameters to be used when 
fitting the model.}

\item{\code{gene_sets_list}}{list of Genesets to be tested for enrichment.}

\item{\code{sea_params}}{SEAparams object with the parameters to be used 
by SEA.}

\item{\code{gsea_params}}{GSEAparams object with the parameters to be used 
by GSEA.}
}}
\examples{
## Lets create a basic IGSAinput object.
## First create a expression matrix.
maData <- matrix(rnorm(10000),ncol=4);
rownames(maData) <- 1:nrow(maData); # It must have rownames (gene names).
maExprData <- new("MAList",list(M=maData));

## Now lets create the FitOptions object.
myFOpts <- FitOptions(c("Cond1", "Cond1", "Cond2", "Cond2"));

## Finally lets create the Genesets to test for enrichment.
myGs1 <- Geneset(id="fakeId1", name="fakeName1", genes=as.character(1:10));
myGs2 <- Geneset(id="fakeId2", name="fakeName2", genes=as.character(7:15));
myGSs <- Genesets(name="myGenesets", gene_sets=list(myGs1, myGs2),
is_GO=FALSE);

## And now we can create our IGSAinput ready for MIGSA.
igsaInput <- IGSAinput(name="igsaInput", expr_data=maExprData, 
fit_options=myFOpts, gene_sets_list=list(myGSs));

}

