% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGSAinput-common.R
\docType{methods}
\name{IGSAinput-common}
\alias{IGSAinput-common}
\alias{summary,IGSAinput-method}
\alias{summary.IGSAinput}
\title{IGSAinput exploratory functions}
\usage{
\method{summary}{IGSAinput}(object, ...)
}
\arguments{
\item{object}{IGSAinput object.}

\item{...}{not in use.}
}
\value{
a summary of the object.
}
\description{
Several R base overwritten functions to manipulate a IGSAinput object.
}
\examples{
## Lets create a basic IGSAinput object.
## First create a expression matrix.
maData <- matrix(rnorm(10000),ncol=4);
rownames(maData) <- 1:nrow(maData); # It must have rownames (gene names).
maExprData <- new("MAList",list(M=maData));

## Now lets create the FitOptions object.
myFOpts <- FitOptions(c("Cond1", "Cond1", "Cond2", "Cond2"));

## And now we can create our IGSAinput ready for MIGSA.
igsaInput <- IGSAinput(name="myIgsaInput", expr_data=maExprData, 
fit_options=myFOpts);
summary(igsaInput);

}

