% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGSAinput-getDEGenes.R
\docType{methods}
\name{getDEGenes}
\alias{getDEGenes}
\alias{getDEGenes,IGSAinput-method}
\alias{getDEGenes,IGSAinput}
\title{Calculate differentialy expressed genes of an IGSAinput object}
\usage{
getDEGenes(igsaInput)

\S4method{getDEGenes}{IGSAinput}(igsaInput)
}
\arguments{
\item{igsaInput}{a valid IGSAinput object.}
}
\value{
A IGSAinput object with the updated slots and DE genes calculated.
}
\description{
\code{getDEGenes} calculates the differentialy expressed genes of an 
IGSAinput object using the expression matrix, the FitOtions and the 
SEAparams..
}
\examples{
## Lets create a basic IGSAinput object.
## First create a expression matrix.
maData <- matrix(rnorm(10000),ncol=4);
rownames(maData) <- 1:nrow(maData); # It must have rownames (gene names).
maExprData <- new("MAList",list(M=maData));

## Now lets create the FitOptions object.
myFOpts <- FitOptions(c("Cond1", "Cond1", "Cond2", "Cond2"));

## Lets create our IGSAinput object.
myIgsaInput <- IGSAinput(name="myIgsaInput", expr_data=maExprData, 
fit_options=myFOpts);

## And check how many differentialy expressed genes does it get with actual
## parameters.
aux <- getDEGenes(myIgsaInput);

}
\seealso{
\code{\link{IGSAinput-class}}
}
