% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAinput-common.R
\docType{methods}
\name{MIGSAinput-common}
\alias{MIGSAinput-common}
\alias{summary,MIGSAinput-method}
\alias{summary.MIGSAinput}
\title{MIGSAinput exploratory functions}
\usage{
\method{summary}{MIGSAinput}(object, ...)
}
\arguments{
\item{object}{MIGSAinput object.}

\item{...}{not in use.}
}
\value{
a summary of the object.
}
\description{
Several R base overwritten functions to manipulate a MIGSAinput object.
}
\examples{
## Lets create a basic MIGSAinput object with one IGSAinput object.

## First create an expression matrix.
maData <- matrix(rnorm(10000),ncol=4);
rownames(maData) <- 1:nrow(maData); # It must have rownames (gene names).
maExprData <- new("MAList",list(M=maData));

## Now lets create the FitOptions object.
myFOpts <- FitOptions(c("Cond1", "Cond1", "Cond2", "Cond2"));

## Lets create our IGSAinput ready for MIGSA.
igsaInput <- IGSAinput(name="igsaInput", expr_data=maExprData, 
fit_options=myFOpts);

## Finally lets create out MIGSAinput object.
migsaInput <- MIGSAinput(experiments=list(igsaInput));
summary(migsaInput);

}

