% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAinput-getterSetters.R
\docType{methods}
\name{MIGSAinput-getterSetters}
\alias{MIGSAinput-getterSetters}
\alias{bpParam}
\alias{bpParam,MIGSAinput-method}
\alias{bpParam<-}
\alias{bpParam<-,MIGSAinput-method}
\alias{experiments}
\alias{experiments,MIGSAinput-method}
\alias{experiments<-}
\alias{experiments<-,MIGSAinput-method}
\alias{geneSetsList,MIGSAinput-method}
\alias{geneSetsList<-,MIGSAinput-method}
\title{Accessors for MIGSAinput class.}
\usage{
"MIGSAinput-getterSetters"(object)

experiments(object)

\S4method{experiments}{MIGSAinput}(object)

experiments(object) <- value

\S4method{experiments}{MIGSAinput}(object) <- value

bpParam(object)

\S4method{bpParam}{MIGSAinput}(object)

bpParam(object) <- value

\S4method{bpParam}{MIGSAinput}(object) <- value

\S4method{geneSetsList}{MIGSAinput}(object)

\S4method{geneSetsList}{MIGSAinput}(object) <- value
}
\arguments{
\item{object}{MIGSAinput object.}

\item{value}{value to replace in the slot.}
}
\value{
modified object or desired slot.
}
\description{
Getters and setters functions to access MIGSAinput object slots.
}
\examples{
## Lets create an empty MIGSAinput object and access its slots.
migsaInput <- MIGSAinput();

## Lets get out migsaInput values.
bpParam(migsaInput);
experiments(migsaInput);
geneSetsList(migsaInput);

}

