% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAPackage.R
\docType{package}
\name{MIGSAPackage}
\alias{MIGSAPackage}
\alias{MIGSAPackage-package}
\title{MIGSA package}
\description{
The MIGSA package permits to develop an extensive gene set enrichment 
analysis over multiple datasets. It provides a common gene expression 
analytic framework that grants a comprehensive and coherent analysis. Only a 
minimal user parameter setting is required, since the best SEA/GSEA 
alternatives are integrated, i.e. \pkg{dEnricher} and \pkg{mGSZ} respectively.
MIGSA's greatest strengths are its big data tools, as it provides several 
functions to explore, analyze and visualize its results in order to 
facilitate the data mining task over huge information sources.
}
\author{
Juan C. Rodriguez \email{jcrodriguez@bdmg.com.ar}, Cristobal Fresno 
\email{cfresno@bdmg.com.ar}, Andrea S. Llera \email{allera@leloir.org.ar} 
and Elmer A. Fernandez \email{efernandez@bdmg.com.ar}
}
\references{
\enumerate{
\item Rodriguez, J. C., Gonzalez, G. A., Fresno, C., Llera, A. S., & 
Fernandez, E. A. (2016). Improving information retrieval in functional 
analysis. Computers in Biology and Medicine, 79, 10-20.
}
}
\keyword{Analysis}
\keyword{Functional}

