% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAres-class.R
\docType{class}
\name{MIGSAres-class}
\alias{MIGSAres-class}
\title{MIGSAres S4 class implementation in R}
\description{
This S4 class represents the results of one MIGSA execution in R.
}
\seealso{
\code{\link{MIGSA}}

\code{\link{getAdditionalInfo}}

\code{\link{MIGSAres-common}}

\code{\link{MIGSAres-genes}}

\code{\link{MIGSAres-GOanalysis}}

\code{\link{MIGSAres-plots}}

\code{\link{setEnrCutoff}}

\code{\link{summary}}
}
