% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAres-filterByGenes.R
\docType{methods}
\name{filterByGenes}
\alias{filterByGenes}
\alias{filterByGenes,MIGSAres,character-method}
\title{Filters MIGSAres by selected genes}
\usage{
filterByGenes(migsaRes, genes)

\S4method{filterByGenes}{MIGSAres,character}(migsaRes, genes)
}
\arguments{
\item{migsaRes}{MIGSAres object.}

\item{genes}{character vector of the interest genes for MIGSAres filtering.}
}
\value{
MIGSAres object containing only the gene sets in which genes 
contributed to enrichment.
}
\description{
\code{filterByGenes} returns a MIGSAres object with only the gene sets which 
resulted enriched by at least one gene from a provided list.
}
\examples{
data(migsaRes);

## First lets set a cutoff of 0.1
migsaResWCoff <- setEnrCutoff(migsaRes, 0.1);

## Suppose we are interested in studying these genes:
intGenes <- c("g10", "g91", "g388", "g742", "g874");
migsaResIntGenes <- filterByGenes(migsaResWCoff, intGenes);

## Now in migsaResIntGenes we have the MIGSA results of the gene sets in 
## which at least one gene of our list contributed to enrich.
migsaResIntGenes

}
\seealso{
\code{\link{setEnrCutoff}} and \code{\link{genesInSets}}
}

