% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAres-genesBarplot.R
\docType{methods}
\name{genesBarplot}
\alias{genesBarplot}
\alias{genesBarplot,MIGSAres-method}
\title{Genes barplot of the number of gene sets in which each enrich}
\usage{
genesBarplot(migsaRes, ...)

\S4method{genesBarplot}{MIGSAres}(migsaRes, enrFilter = 0, gsFilter = 0)
}
\arguments{
\item{migsaRes}{MIGSAres object.}

\item{...}{not in use.}

\item{enrFilter}{numeric. Keep gene sets enriched in at least enrFilter 
experiments.}

\item{gsFilter}{numeric. Keep genes enriched in at least gsFilter gene sets.}
}
\value{
ggplot object used as graphic.
}
\description{
\code{genesBarplot} generates a barplot of the number of gene sets in which 
each gene contributed to enrich. Each gene set counts 1 regardless if it was 
enriched in many experiments. x-axis each gene, y-axis number of gene sets 
in which it contributed to enrich.
}
\examples{
data(migsaRes);

## First lets set a cutoff of 0.01
migsaResWCoff <- setEnrCutoff(migsaRes, 0.01);

## Lets check what genes contributed to enrich the highest number of gene 
## sets (in more than one gene set).
genesBarplot(migsaResWCoff, gsFilter=1);

## Moreover we can keep gene sets which where enriched in more than 
## enrFilter experiments. To do this, we can use the enrFilter parameter.

}
\seealso{
\code{\link{setEnrCutoff}}
}

