% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAres-geneSetBarplot.R
\docType{methods}
\name{geneSetBarplot}
\alias{geneSetBarplot}
\alias{geneSetBarplot,MIGSAres-method}
\title{Gene set barplot
Gene set barplot of the number of experiments in which each enrich}
\usage{
geneSetBarplot(migsaRes, ...)

\S4method{geneSetBarplot}{MIGSAres}(migsaRes, enrFilter = 0)
}
\arguments{
\item{migsaRes}{MIGSAres object.}

\item{...}{not in use.}

\item{enrFilter}{numeric. Keep gene sets enriched in at least enrFilter 
experiments.}
}
\value{
ggplot object used as graphic.
}
\description{
\code{geneSetBarplot} generates a barplot of the number of experiments in 
which each gene set was enriched. x-axis each gene set, y-axis times it was 
enriched (0 to #experiments).
}
\examples{
data(migsaRes);

## First lets set a cutoff of 0.1
migsaResWCoff <- setEnrCutoff(migsaRes, 0.1);

## Lets check in how many experiments each gene set was enriched (in more 
## than one experiment).
geneSetBarplot(migsaResWCoff, enrFilter=1);

}
\seealso{
\code{\link{setEnrCutoff}}
}

