% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAres-genesHeatmap.R
\docType{methods}
\name{genesHeatmap}
\alias{genesHeatmap}
\alias{genesHeatmap,MIGSAres-method}
\title{Plots genes heatmap}
\usage{
genesHeatmap(migsaRes, ...)

\S4method{genesHeatmap}{MIGSAres}(migsaRes, enrFilter = 0, gsFilter = 0,
...)
}
\arguments{
\item{migsaRes}{MIGSAres object.}

\item{...}{additional parameters passed to heatmap.2 function.}

\item{enrFilter}{numeric. Keep gene sets enriched in at least enrFilter 
experiments.}

\item{gsFilter}{numeric. Keep genes enriched in at least gsFilter gene sets.}
}
\value{
list returned by heatmap.2 function (plotted data).
}
\description{
\code{genesHeatmap} plots a heatmap with the number of experiments in which 
each gene contributed to enrich each gene set.
}
\examples{
data(migsaRes);

## First lets set a cutoff of 0.1
migsaResWCoff <- setEnrCutoff(migsaRes, 0.1);

## Lets check what genes contributed to enrich the highest number of gene 
## sets (in more than one gene set).
genesHeatmap(migsaResWCoff, gsFilter=1);

## Moreover we can keep gene sets which where enriched in more than 
## enrFilter experiments. To do this, we can use the enrFilter parameter.

}
\seealso{
\code{\link{setEnrCutoff}}
}

