% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAres-genesInSets.R
\docType{methods}
\name{genesInSets}
\alias{genesInSets}
\alias{genesInSets,MIGSAres}
\alias{genesInSets,MIGSAres-method}
\title{Get the genes that contributed to enrichment}
\usage{
genesInSets(migsaRes)

\S4method{genesInSets}{MIGSAres}(migsaRes)
}
\arguments{
\item{migsaRes}{MIGSAres object.}
}
\value{
data.frame with the number of experiments in which each gene 
contributed to enrich each gene set.
}
\description{
\code{genesInSets} returns a data.frame with gene sets as rows, genes as 
columns, and as value the number of experiments in which each gene 
contributed to enrich each gene set. 
If it was enriched only by SEA then it returns the genes that contributed in 
SEA. If it was enriched only by GSEA then it returns the genes that 
contributed in GSEA. If it was enriched by both then it returns the genes 
that contributed in both.
}
\examples{
data(migsaRes);

## First lets set a cutoff of 0.1
migsaResWCoff <- setEnrCutoff(migsaRes, 0.1);
gInSets <- genesInSets(migsaResWCoff);
class(gInSets); # matrix

## Now we can do stuff as check which genes enriched a gene set in all (two) 
## experiments.
gInSets[rowSums(gInSets==2) > 0, colSums(gInSets==2) > 0];

}
\seealso{
\code{\link{setEnrCutoff}}
}

