% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAres-genesManipulation.R
\docType{methods}
\name{MIGSAres-genes}
\alias{MIGSAres-genes}
\alias{genesInSets}
\alias{genesInSets,MIGSAres-method}
\alias{filterByGenes}
\alias{filterByGenes,MIGSAres,character-method}
\title{Get or edit MIGSAres with/by genes that contributed to enrichment}
\usage{
"MIGSAres-genes"(migsaRes)

genesInSets(migsaRes)

\S4method{genesInSets}{MIGSAres}(migsaRes)

filterByGenes(migsaRes, genes)

\S4method{filterByGenes}{MIGSAres,character}(migsaRes, genes)
}
\arguments{
\item{migsaRes}{MIGSAres object.}

\item{genes}{character vector of the interest genes for MIGSAres filtering.}
}
\value{
If genesInSets: a data.frame with the number of experiments in which 
each gene contributed to enrich each gene set. If filterByGenes: A MIGSAres 
object containing only the gene sets in which provided genes contributed to 
enrichment.
}
\description{
\code{genesInSets} returns a data.frame with gene sets as rows, genes as 
columns, and as value the number of experiments in which each gene 
contributed to enrich each gene set. 
If it was enriched only by SEA then it returns the genes that contributed in 
SEA. If it was enriched only by GSEA then it returns the genes that 
contributed in GSEA. If it was enriched by both then it returns the genes 
that contributed in both.
\code{filterByGenes} returns a MIGSAres object with only the gene sets which 
resulted enriched by at least one gene from a provided list.
}
\examples{
data(migsaRes);

###### genesInSets
## First lets set a cutoff of 0.1
migsaResWCoff <- setEnrCutoff(migsaRes, 0.1);
gInSets <- genesInSets(migsaResWCoff);
class(gInSets); # matrix

## Now we can do stuff as check which genes enriched a gene set in all (two) 
## experiments.
gInSets[rowSums(gInSets==2) > 0, colSums(gInSets==2) > 0];

###### filterByGenes
## Suppose we are interested in studying these genes:
intGenes <- c("g10", "g91", "g388", "g742", "g874");
migsaResIntGenes <- filterByGenes(migsaResWCoff, intGenes);

## Now in migsaResIntGenes we have the MIGSA results of the gene sets in 
## which at least one gene of our list contributed to enrich.
migsaResIntGenes

}
\seealso{
\code{\link{setEnrCutoff}}
}
