% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoAnalysis-getHeights.R
\docType{methods}
\name{getHeights}
\alias{getHeights}
\alias{getHeights,character-method}
\title{Gets the Gene Ontology depths of specified ids}
\usage{
getHeights(ids, ...)

\S4method{getHeights}{character}(ids, minHeight = TRUE)
}
\arguments{
\item{ids}{character vector indicating the queried GO ids.}

\item{...}{not in use.}

\item{minHeight}{logical indicating if the minimum or maximum height must be 
calculated. If it is FALSE then the longest path to the root is calculated, 
otherwise, the shortest path.}
}
\value{
list with each term height.
}
\description{
\code{getHeights} returns the heights of given a list of ids (GO IDs).
}
\examples{
## Lets load breast cancer results.
data(bcMigsaRes);

## And get the first 40 Gene Ontology gene sets IDs.
goIds <- bcMigsaRes[bcMigsaRes$GS_Name \%in\% c("BP", "CC", "MF"), "id"];
fst40goIds <- goIds[1:40,];

## And lets get the heights in the GO tree structure.
getHeights(fst40goIds);

}

