% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoAnalysis-migsaGoTree.R
\docType{methods}
\name{migsaGoTree}
\alias{migsaGoTree}
\alias{migsaGoTree,MIGSAres}
\alias{migsaGoTree,MIGSAres-method}
\title{Plots the GO tree/s}
\usage{
migsaGoTree(migsaRes)

\S4method{migsaGoTree}{MIGSAres}(migsaRes)
}
\arguments{
\item{migsaRes}{MIGSAres object. It must contain at least one GO Geneset.}
}
\value{
list with the used data to plot.
}
\description{
\code{migsaGoTree} Plots the GO tree/s present in migsaRes,
}
\examples{
## Lets load breast cancer results.
data(bcMigsaRes);

## And get the first 40 Gene Ontology gene sets results from CC.
goRes <- bcMigsaRes[bcMigsaRes$GS_Name == "CC",];
fst40goRes <- goRes[1:40,];

## And lets plot the results GO trees.
aux <- migsaGoTree(fst40goRes);

}

