% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAres-migsaHeatmap.R
\docType{methods}
\name{migsaHeatmap}
\alias{migsaHeatmap}
\alias{migsaHeatmap,MIGSAres}
\alias{migsaHeatmap,MIGSAres-method}
\title{Plots enrichment heatmap}
\usage{
migsaHeatmap(migsaRes, ...)

\S4method{migsaHeatmap}{MIGSAres}(migsaRes, flevel = 0, penrich = 0,
col.dist = "jaccard", row.dist = col.dist, ...)
}
\arguments{
\item{migsaRes}{MIGSAres object.}

\item{...}{additional parameters passed to heatmap.2 function.}

\item{flevel}{numeric. Keeps gene sets enriched in at least flevel 
experiments.}

\item{penrich}{numeric. Keeps experiments which enriched at least penrich 
gene sets.}

\item{col.dist}{character. Distance algorithm to be used in columns, passed 
to vegdist function.}

\item{row.dist}{character. Distance algorithm to be used in rows, passed to 
vegdist function.}
}
\value{
list returned by heatmap.2 function (plotted data).
}
\description{
\code{migsaHeatmap} plots the enrichment heatmap of the MIGSAres object.
}
\examples{
data(migsaRes);

## First lets set a cutoff of 0.1
migsaResWCoff <- setEnrCutoff(migsaRes, 0.1);

## Lets visually check enriched gene sets shared between experiments.
migsaHeatmap(migsaResWCoff);

}
\seealso{
\code{\link{setEnrCutoff}}
}

