% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAres-plotting.R
\docType{methods}
\name{MIGSAres-plots}
\alias{MIGSAres-plots}
\alias{genesHeatmap}
\alias{genesHeatmap,MIGSAres-method}
\alias{genesBarplot}
\alias{genesBarplot,MIGSAres-method}
\alias{migsaHeatmap}
\alias{migsaHeatmap,MIGSAres-method}
\alias{migsaHeatmap,matrix-method}
\alias{geneSetBarplot}
\alias{geneSetBarplot,MIGSAres-method}
\title{MIGSAres plots}
\usage{
"MIGSAres-plots"(migsaRes)

genesHeatmap(migsaRes, ...)

\S4method{genesHeatmap}{MIGSAres}(migsaRes, enrFilter = 0, gsFilter = 0,
colPal = c("white", "red"), col.dist = NA, row.dist = NA, layout = NA,
dendrogram = "row")

genesBarplot(migsaRes, ...)

\S4method{genesBarplot}{MIGSAres}(migsaRes, enrFilter = 0, gsFilter = 0)

migsaHeatmap(migsaRes, ...)

\S4method{migsaHeatmap}{MIGSAres}(migsaRes, enrFilter = 0, expFilter = 0,
categories = list(), categLabels = TRUE, colPal = c("white", "red"),
col.dist = NA, row.dist = NA, layout = NA, remove0Rows = TRUE,
breaks = NA)

\S4method{migsaHeatmap}{matrix}(migsaRes, categories = list(),
categLabels = TRUE, colPal = c("white", "red"), col.dist = NA,
row.dist = NA, layout = NA)

geneSetBarplot(migsaRes, ...)

\S4method{geneSetBarplot}{MIGSAres}(migsaRes, enrFilter = 0)
}
\arguments{
\item{migsaRes}{MIGSAres object.}

\item{...}{not in use.}

\item{enrFilter}{numeric. Keep gene sets enriched in at least enrFilter 
experiments.}

\item{gsFilter}{numeric. Keep genes enriched in at least gsFilter gene sets.}

\item{colPal}{vector. Character vector of two colors, first value will 
represent FALSE/1 on heatmap, and second value TRUE/0.}

\item{col.dist}{character. Distance algorithm to be used in columns, passed 
to vegdist function. If migsaRes has cutoff then default is jaccard, else, 
default is euclidean.}

\item{row.dist}{character. Distance algorithm to be used in rows, passed to 
vegdist function. If migsaRes has cutoff then default is jaccard, else, 
default is euclidean.}

\item{layout}{matrix. The layout used for heatmaps.}

\item{dendrogram}{character. Wheter to plot "row", "col", "none" or "both"
dendrograms.}

\item{expFilter}{numeric. Keep experiments which enriched at least expFilter 
gene sets.}

\item{categories}{list. List of character vectors, each vector must have the
same length as the number of experiments (ncol(migsaRes)-3). It will plot 
for each category/column a color representing its category.}

\item{categLabels}{logical. Indicates if labels should be plotted for each 
category.}

\item{remove0Rows}{logical. Whether remove gene sets that are not enriched in 
any experiment.}

\item{breaks}{numeric. If migsaRes does not have cutoff then break P-values 
in breaks intervals.}
}
\value{
In heatmap functions: A list returned by heatmap.2 function 
(plotted data). In other functions: A ggplot object used as graphic.
}
\description{
\code{genesHeatmap} plots a heatmap with the number of experiments in which 
each gene contributed to enrich each gene set.
\code{genesBarplot} generates a barplot of the number of gene sets in which 
each gene contributed to enrich. Each gene set counts 1 regardless if it was 
enriched in many experiments. x-axis each gene, y-axis number of gene sets 
in which it contributed to enrich.
\code{migsaHeatmap} plots the enrichment heatmap of the MIGSAres object.
\code{geneSetBarplot} generates a barplot of the number of experiments in 
which each gene set was enriched. x-axis each gene set, y-axis times it was 
enriched (0 to #experiments).
}
\examples{
data(migsaRes);
## For this example lets work with the first 50 results
migsaRes <- migsaRes[1:50,];

#### genesHeatmap
## First lets set a cutoff of 0.1
migsaResWCoff <- setEnrCutoff(migsaRes, 0.1);

## Lets check what genes contributed to enrich the highest number of gene 
## sets (in more than one gene set).
genesHeatmap(migsaResWCoff, gsFilter=1);

## Moreover we can keep gene sets which where enriched in more than 
## enrFilter experiments. To do this, we can use the enrFilter parameter.

#### genesBarplot
## Lets set a cutoff of 0.01
migsaResWCoff <- setEnrCutoff(migsaRes, 0.01);

## Lets check what genes contributed to enrich the highest number of gene 
## sets (in more than one gene set).
genesBarplot(migsaResWCoff, gsFilter=1);

## Moreover we can keep gene sets which where enriched in more than 
## enrFilter experiments. To do this, we can use the enrFilter parameter.

#### migsaHeatmap
## Lets set a cutoff of 0.1
migsaResWCoff <- setEnrCutoff(migsaRes, 0.1);

## Lets visually check enriched gene sets shared between experiments.
migsaHeatmap(migsaResWCoff);

#### geneSetBarplot
## Lets set a cutoff of 0.1
migsaResWCoff <- setEnrCutoff(migsaRes, 0.1);

## Lets check in how many experiments each gene set was enriched (in more 
## than one experiment).
geneSetBarplot(migsaResWCoff, enrFilter=1);

}
\seealso{
\code{\link{setEnrCutoff}}
}
