% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAres-setEnrCutoff.R
\docType{methods}
\name{setEnrCutoff}
\alias{setEnrCutoff}
\alias{setEnrCutoff,MIGSAres,numeric-method}
\title{Set enrichment cutoff for a MIGSAres object}
\usage{
setEnrCutoff(object, newEnrCutoff)

\S4method{setEnrCutoff}{MIGSAres,numeric}(object, newEnrCutoff)
}
\arguments{
\item{object}{a MIGSAres object.}

\item{newEnrCutoff}{numeric value in range [0,1] or NA to set the new
enrichment cutoff. If NA then enrichment cutoff is unset for object.}
}
\value{
A MIGSAres object with the updated slots.
}
\description{
\code{setEnrCutoff} sets the enrichment cutoff value for a MIGSAres object in
order to get detailed enrichment results.
}
\examples{
data(migsaRes);
## After executing MIGSA, the MIGSAres object does not have an enrichment 
## cutoff set, so it will be shown as follows:
head(migsaRes);

## Now lets set an enrichment cutoff, and then show again the object.
migsaResWCoff <- setEnrCutoff(migsaRes, 0.01);
head(migsaResWCoff);

## So now we can do stuff like check how many gene sets are enriched in both 
## experiments, in each one, etc. Moreover now we can do different MIGSAres 
## plots available in MIGSA package.
table(migsaResWCoff[, c("igsaInput1", "igsaInput2")]);

## And with different cutoffs.
migsaResWCoff <- setEnrCutoff(migsaRes, 0.1);
table(migsaResWCoff[, c("igsaInput1", "igsaInput2")]);

}
