% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-migsaRes.R
\docType{data}
\name{migsaRes}
\alias{migsaRes}
\title{A MIGSA results object (MIGSAres) example}
\format{A MIGSAres object}
\source{
see \code{\link{MIGSA}}
}
\description{
A MIGSA results object (MIGSAres) obtained using the code present as example 
in \link{MIGSA} man page.
}
