% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAparams-common.R
\docType{methods}
\name{SEAparams-common}
\alias{SEAparams-common}
\alias{summary,SEAparams-method}
\alias{summary.SEAparams}
\title{SEAparams exploratory functions}
\usage{
\method{summary}{SEAparams}(object, ...)
}
\arguments{
\item{object}{SEAparams object.}

\item{...}{not in use.}
}
\value{
a summary of the object.
}
\description{
Several R base overwritten functions to manipulate a SEAparams object.
}
\examples{
seaParams <- SEAparams();
summary(seaParams);

}

