% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets-asGenesets.R
\docType{methods}
\name{as.Genesets}
\alias{as.Genesets}
\alias{as.Genesets,list-method}
\title{Creates a Genesets from a list}
\usage{
as.Genesets(x, ...)

\S4method{as.Genesets}{list}(x, name = "no_name")
}
\arguments{
\item{x}{list of character vectors which are the genes corresponding to each 
Geneset. The list must have names (unique).}

\item{...}{not in use.}

\item{name}{character indicating the name of these gene sets e.g. "KEGG" 
(default: "no_name").}
}
\value{
Genesets object.
}
\description{
\code{as.Genesets} creates a Genesets object from the data present in a list.
Each element will parse to a Geneset. For each list element, its name will be 
the Geneset id, and the content are the genes.
}
\examples{
## Lets create a list with three manually created gene sets and load it as a 
## Genesets object.
myGs1 <- as.character(1:10);
myGs2 <- as.character(15:21);
myGs3 <- as.character(25:30);
myGssList <- list(myGs1, myGs2, myGs3);
names(myGssList) <- c("myGs1", "myGs2", "myGs3");
myGss <- as.Genesets(myGssList, name="myGeneSets");

}

