% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets-enrichr.R
\docType{methods}
\name{Genesets-enrichr}
\alias{Genesets-enrichr}
\alias{enrichrGeneSets}
\alias{enrichrGeneSets,character-method}
\alias{enrichrGeneSets,ANY-method}
\alias{downloadEnrichrGeneSets}
\alias{downloadEnrichrGeneSets,character-method}
\alias{downloadEnrichrGeneSets,ANY-method}
\title{List and download gene sets from enrichr database}
\usage{
"Genesets-enrichr"(pattern)

enrichrGeneSets(pattern = ".*")

\S4method{enrichrGeneSets}{ANY}(pattern = ".*")

downloadEnrichrGeneSets(geneSetNames, ...)

\S4method{downloadEnrichrGeneSets}{character}(geneSetNames,
deleteMultipleEntrez = !FALSE)
}
\arguments{
\item{pattern}{character indicating a pattern to filter the database names.}

\item{geneSetNames}{list of characters with the names of the gene sets to 
download. Must be listed at \code{\link{enrichrGeneSets}}.}

\item{...}{not in use.}

\item{deleteMultipleEntrez}{logical indicating if multiple Entrez IDs should 
be deleted or repeated.
Note: Enrichr uses Gene Symbol, if org.Hs.eg translates it into several 
Entrez IDs then if deleteMultipleEntrez == FALSE all Entrez are removed else 
all are included.}
}
\value{
enrichrGeneSets: character with present database names.
downloadEnrichrGeneSets: list of GeneSetCollection objects (Genes are with 
their EntrezGene ID).
}
\description{
\code{enrichrGeneSets} lists the database names present at enrichr.
\code{downloadEnrichrGeneSets} creates a list of GeneSetCollection 
objects downloading the specified ones from enrichr website
(http://amp.pharm.mssm.edu /Enrichr/).
}
\examples{
## Lets list all the gene sets that can be downloaded from Enichr website.
enrichrGeneSets();

\dontrun{
## Now lets list only the gene sets that have BioCarta in their names 
## (different BioCarta versions).
enrichrGeneSets("BioCarta");
}

## And lets download the latest BioCarta gene sets from Enrichr.
## Make sure you use the same names as listed with enrichrGeneSets() .
biocartaGSs <- downloadEnrichrGeneSets(c("BioCarta_2015"));

}
\seealso{
\code{\link{as.Genesets}}

\code{\link{geneSetsFromFile}}

\code{\link{loadGo}}
}
