% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets-loadGo.R
\docType{methods}
\name{loadGo}
\alias{loadGo}
\alias{loadGo,character-method}
\title{Creates a Genesets object using the Gene Ontology data base}
\usage{
loadGo(ontology)

\S4method{loadGo}{character}(ontology)
}
\arguments{
\item{ontology}{character indicating which ontology must be loaded. 
Must be one of BP, MF or CC.}
}
\value{
Genesets object including the desired Gene Ontology gene sets.
}
\description{
\code{loadGo} creates a Genesets object from the data present at the Gene 
Ontology data base.
}
\examples{
## Lets load the Cellular Components gene sets from the Gene Ontology.
ccGSets <- loadGo("CC");

}

