% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAres-GoAnalysis.R
\docType{methods}
\name{MIGSAres-GOanalysis}
\alias{MIGSAres-GOanalysis}
\alias{getHeights}
\alias{getHeights,character-method}
\alias{migsaGoTree}
\alias{migsaGoTree,MIGSAres-method}
\title{Explore Gene Ontology gene sets in MIGSAres}
\usage{
"MIGSAres-GOanalysis"(migsaRes)

migsaGoTree(migsaRes)

\S4method{migsaGoTree}{MIGSAres}(migsaRes)

getHeights(ids, ...)

\S4method{getHeights}{character}(ids, minHeight = TRUE)
}
\arguments{
\item{migsaRes}{MIGSAres object. It must contain at least one GO gene set.}

\item{ids}{character vector indicating the queried GO ids.}

\item{...}{not in use.}

\item{minHeight}{logical indicating if the minimum or maximum height must be 
calculated. If it is FALSE then the longest path to the root is calculated, 
otherwise, the shortest path.}
}
\value{
If migsaGoTree: A list with the used data to plot. If getHeights: A 
list with each term height.
}
\description{
\code{migsaGoTree} plots the GO tree/s present in migsaRes.
\code{getHeights} returns the heights of given a list of ids (GO IDs).
}
\examples{
## Lets load breast cancer results.
data(bcMigsaRes);

###### migsaGoTree
## Get the first 40 Gene Ontology gene sets results from CC.
goRes <- bcMigsaRes[bcMigsaRes$GS_Name == "CC",];
fst40goRes <- goRes[1:40,];

## And lets plot the results GO trees.
\dontrun{
aux <- migsaGoTree(fst40goRes);
}

###### getHeights
## Get the first 40 Gene Ontology gene sets IDs.
goIds <- bcMigsaRes[bcMigsaRes$GS_Name \%in\% c("BP", "CC", "MF"), "id"];
fst40goIds <- goIds[1:40,];

\dontrun{
## And lets get the heights in the GO tree structure.
getHeights(fst40goIds);
}

}

