% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAres-plotting.R
\docType{methods}
\name{MIGSAres-plots}
\alias{MIGSAres-plots}
\alias{geneSetBarplot}
\alias{geneSetBarplot,MIGSAres-method}
\alias{genesBarplot}
\alias{genesBarplot,MIGSAres-method}
\alias{genesHeatmap}
\alias{genesHeatmap,MIGSAres-method}
\alias{migsaHeatmap}
\alias{migsaHeatmap,MIGSAres-method}
\title{MIGSAres plots}
\usage{
"MIGSAres-plots"(migsaRes)

genesHeatmap(migsaRes, ...)

\S4method{genesHeatmap}{MIGSAres}(migsaRes, enrFilter = 0, gsFilter = 0,
...)

genesBarplot(migsaRes, ...)

\S4method{genesBarplot}{MIGSAres}(migsaRes, enrFilter = 0, gsFilter = 0)

migsaHeatmap(migsaRes, ...)

\S4method{migsaHeatmap}{MIGSAres}(migsaRes, enrFilter = 0, expFilter = 0,
col.dist = "jaccard", row.dist = col.dist, ...)

geneSetBarplot(migsaRes, ...)

\S4method{geneSetBarplot}{MIGSAres}(migsaRes, enrFilter = 0)
}
\arguments{
\item{migsaRes}{MIGSAres object.}

\item{...}{In heatmap functions, ... are additional parameters passed to 
heatmap.2 function. In other functions it is not in use.}

\item{enrFilter}{numeric. Keep gene sets enriched in at least enrFilter 
experiments.}

\item{gsFilter}{numeric. Keep genes enriched in at least gsFilter gene sets.}

\item{expFilter}{numeric. Keep experiments which enriched at least expFilter 
gene sets.}

\item{col.dist}{character. Distance algorithm to be used in columns, passed 
to vegdist function.}

\item{row.dist}{character. Distance algorithm to be used in rows, passed to 
vegdist function.}
}
\value{
In heatmap functions: A list returned by heatmap.2 function 
(plotted data). In other functions: A ggplot object used as graphic.
}
\description{
\code{genesHeatmap} plots a heatmap with the number of experiments in which 
each gene contributed to enrich each gene set.
\code{genesBarplot} generates a barplot of the number of gene sets in which 
each gene contributed to enrich. Each gene set counts 1 regardless if it was 
enriched in many experiments. x-axis each gene, y-axis number of gene sets 
in which it contributed to enrich.
\code{migsaHeatmap} plots the enrichment heatmap of the MIGSAres object.
\code{geneSetBarplot} generates a barplot of the number of experiments in 
which each gene set was enriched. x-axis each gene set, y-axis times it was 
enriched (0 to #experiments).
}
\examples{
data(migsaRes);

#### genesHeatmap
## First lets set a cutoff of 0.1
migsaResWCoff <- setEnrCutoff(migsaRes, 0.1);

## Lets check what genes contributed to enrich the highest number of gene 
## sets (in more than one gene set).
genesHeatmap(migsaResWCoff, gsFilter=1);

## Moreover we can keep gene sets which where enriched in more than 
## enrFilter experiments. To do this, we can use the enrFilter parameter.

#### genesBarplot
## Lets set a cutoff of 0.01
migsaResWCoff <- setEnrCutoff(migsaRes, 0.01);

## Lets check what genes contributed to enrich the highest number of gene 
## sets (in more than one gene set).
genesBarplot(migsaResWCoff, gsFilter=1);

## Moreover we can keep gene sets which where enriched in more than 
## enrFilter experiments. To do this, we can use the enrFilter parameter.

#### migsaHeatmap
## Lets set a cutoff of 0.1
migsaResWCoff <- setEnrCutoff(migsaRes, 0.1);

## Lets visually check enriched gene sets shared between experiments.
migsaHeatmap(migsaResWCoff);

#### geneSetBarplot
## Lets set a cutoff of 0.1
migsaResWCoff <- setEnrCutoff(migsaRes, 0.1);

\dontrun{
## Lets check in how many experiments each gene set was enriched (in more 
## than one experiment).
geneSetBarplot(migsaResWCoff, enrFilter=1);
}

}
\seealso{
\code{\link{setEnrCutoff}}
}

